Remove aconto_frequency — use billing_frequency + payment_model
In real Danish electricity billing, there is no separate "aconto frequency."
The billing_frequency (weekly/monthly/quarterly) IS the invoicing cadence,
and payment_model (aconto/post_payment) determines prepaid vs post-paid.

- Drop aconto_frequency from signup and contract tables (V011+V012 migrations)
- Remove daily billing frequency (not a valid Danish market option)
- Aconto invoicing now groups runs by billing_frequency period boundaries
- Add payment_model to signup detail API response and backoffice UI
- Add 15 new unit tests covering all frequency × payment_model combinations
- Fix CustomerBalance int→long for invoice/overdue counts

+++ DataHub.Settlement/src/DataHub.Settlement.Application/Billing/InvoiceModels.cs
@@ -90,8 +90,8 @@ public record CustomerBalance(
     decimal TotalPaid,
     decimal TotalOutstanding,
     decimal TotalOverdue,
-    int InvoiceCount,
-    int OverdueCount);
+    long InvoiceCount,
+    long OverdueCount);
 
 public record CustomerLedgerEntry(
     DateTime Date,
